ultimo commit de la tarde 11 de mayo del 2026

+++ control-backend/src/category/category.service.ts
@@ -1,7 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateCategoryDto } from './create-category.dto';
 import { UpdateCategoryDto } from './update.category.dto';
+import { defaultCategories } from './default-category';
 
 @Injectable()
 export class CategoriesService {
@@ -9,10 +10,46 @@ export class CategoriesService {
 
     // 🔹 Crear categoría para el usuario logueado
     async create(userId: string, dto: CreateCategoryDto) {
+
+        if (dto.parentId) {
+            const parentCategory = await this.prisma.category.findFirst({
+                where: {
+                    id: dto.parentId,
+                    userId,
+                },
+            });
+            if (!parentCategory) {
+                throw new NotFoundException('Parent category not found');
+            }
+        }
+        if (dto.parentId && dto.parentId === dto.name) {
+            throw new BadRequestException('la Subcategoria es la misma que la catgeoria padre');
+        }
+        if (dto.parentId) {
+            const parent = await this.prisma.category.findFirst({
+                where: { id: dto.parentId, userId },
+            });
+
+            if (parent?.type !== dto.type) {
+                throw new BadRequestException('Type mismatch with parent category');
+            }
+        }
+        const exists = await this.prisma.category.findFirst({
+            where: {
+                name: dto.name,
+                userId,
+                parentId: dto.parentId ?? null,
+            },
+        });
+
+        if (exists) {
+            throw new BadRequestException('Category already exists');
+        }
         return this.prisma.category.create({
             data: {
                 name: dto.name,
                 type: dto.type,
+                color: dto.color,
                 ...(dto.parentId ? { parentId: dto.parentId } : {}),
                 userId,
             },
@@ -27,6 +64,7 @@ export class CategoriesService {
             data: {
                 name: dto.name,
                 type: dto.type,
+                color: dto.color,
                 ...(dto.parentId ? { parentId: dto.parentId } : {}),
                 userId,
             },
@@ -55,7 +93,10 @@ export class CategoriesService {
             where: {
                 id,
                 userId,
-            },
+            }, include: {
+                children: true,
+                parent: true,
+            }
         });
 
         if (!category) {
@@ -77,10 +118,99 @@ export class CategoriesService {
 
     // DELETE
     async remove(userId: string, id: string) {
-        await this.findOne(userId, id);
+        const category = await this.findOne(userId, id);
 
+        const isParentCategory = category.children.length > 0;
+        // 🔍 1️⃣ Verificar si es subcategoría o categoría padre
+        const isSubCategory = !!category.parentId;
+
+        // 🔍 2️⃣ Validar transacciones si es SUBCATEGORÍA
+        if (isSubCategory) {
+            const hasTransactions = await this.prisma.transaction.findFirst({
+                where: { subCategoryId: id, userId },
+            });
+
+            if (hasTransactions) {
+                throw new BadRequestException(
+                    'Esta subcategoría tiene transacciones asociadas'
+                );
+            }
+            return await this.prisma.category.delete({
+                where: { id, userId },
+            });
+        }
+        // 🔍 3️⃣ Validar transacciones si es CATEGORÍA PADRE
+        if (isParentCategory) {
+            throw new BadRequestException('Primero elimina las subcategorías');
+        }
+        const hasTransactions = await this.prisma.transaction.findFirst({
+            where: { categoryId: id, userId },
+        });
+
+        if (hasTransactions) {
+            throw new BadRequestException(
+                'Esta categoría tiene transacciones asociadas'
+            );
+        }
+
+        // 🔍 4️⃣ Verificar si tiene subcategorías
+        const hasChildren = await this.prisma.category.findFirst({
+            where: { parentId: id, userId },
+        });
+
+        if (hasChildren) {
+            throw new BadRequestException(
+                'Primero elimina las subcategorías'
+            );
+        }
+        // 🗑️ 5️⃣ eliminar
         return this.prisma.category.delete({
             where: { id },
         });
+    }
+
+    async seedDefaultCategories(userId: string) {
+
+        // 1️⃣ Evitar duplicados (si ya tiene categorías no cargamos nada)
+        const existing = await this.prisma.category.findFirst({
+            where: { userId }
+        });
+
+        if (existing) {
+            throw new BadRequestException("El usuario ya tiene categorías creadas");
+        }
+
+        // 2️⃣ recorrer tipos INCOME / EXPENSE
+        for (const type of ["INCOME", "EXPENSE"] as const) {
+
+            for (const category of defaultCategories[type]) {
+
+                // 3️⃣ crear categoría padre
+                const parent = await this.prisma.category.create({
+                    data: {
+                        name: category.name,
+                        type,
+                        color: category.color,
+                        userId
+                    }
+                });
+
+                // 4️⃣ crear subcategorías hijas
+                for (const sub of category.subcategories) {
+                    await this.prisma.category.create({
+                        data: {
+                            name: sub.name,
+                            type,
+                            color: category.color,
+                            parentId: parent.id,
+                            userId
+                        }
+                    });
+                }
+
+            }
+        }
+
+        return { message: "Estructura inicial cargada correctamente 🚀" };
     }
 }
